pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = "A database management system for document processing and embedding"
authors = ["Vardhan Shorewala <vardhanshorewala@coophive.com>"]
readme = "README.md"
packages = [{include = "src"}]
license = "MIT"
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
keywords = ["database", "document-processing", "embedding"]
maintainers = ["Vardhan Shorewala <vardhanshorewala@coophive.com>"]

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
neo4j = "5.27.0"
certifi = "2025.1.31"
requests = "2.31.0"
requests-unixsocket = "^0.3.0"
psycopg2 = "2.9.9"
numpy = ">=2.1.0"
chromadb = "^1.0.7" 
python-dotenv = "1.0.1"
openai = ">=1.65.2,<2.0.0"
PyPDF2 = "3.0.1"
transformers = "4.48.3"
web3 = "^7.3.1"
flask = "3.0.2"
pyyaml = "6.0.1"
marker-pdf = "^1.6.2"
types-psycopg2 = "^2.9.21.20250318"
types-pyyaml = "^6.0.12.20250402"
sentence-transformers = "^4.1.0"
torch = "^2.5.1"
markdown = "^3.8"
markitdown = {version = "^0.1.1", extras = ["pdf"]}
fastapi = "0.115.9"
uvicorn = "^0.29.0"
pydantic = {extras = ["email"], version = "^2.11.7"}
urllib3 = "<2"

[tool.poetry.group.dev.dependencies]
pytest = "7.4.0"
pytest-cov = "4.1.0"
pytest-mock = "3.11.1"
pytest-env = "0.8.2"
black = "23.7.0"
isort = "5.12.0"
mypy = "1.5.1"
flake8 = "^7.0.0"
autopep8 = "^2.3.2"
autoflake = "^2.3.1"
twine = "^6.1.0"

[build-system]
requires = ["poetry-core>=1.0.0,<2.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
src = "src.main:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "--cov=src --cov-report=xml --cov-report=term"

[tool.black]
line-length = 88
target-version = ["py39", "py310", "py311"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false

[tool.coverage.run]
source = ["src"]
omit = ["tests/*", "**/__init__.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "pass",
    "raise ImportError",
] 